♻️ refactor: search 도메인 파일 이동 - /application/hooks/api/search -> /api/search - useGetUserFindMemes(회원이 찾는 밈 API) 삭제(dead code)
