+++ src/application/hooks/api/index.ts
@@ -1,2 +1,1 @@
 export * from "./account";
-export * from "./search";
