+++ src/pages/explore/tags/[tagId].tsx
@@ -1,8 +1,8 @@
 import { dehydrate, QueryClient } from "@tanstack/react-query";
 import type { GetStaticPaths, GetStaticProps, NextPage } from "next";
 
+import { useGetMemesByTag } from "@/api/search";
 import { useGetTagInfo } from "@/api/tag";
-import { prefetchMemesByTag } from "@/application/hooks";
 import { ExplorePageNavigation } from "@/common/components/Navigation";
 import { NextSeo } from "@/common/components/NextSeo";
 import { PullToRefresh } from "@/common/components/PullToRefresh";
@@ -61,7 +61,7 @@ export const getStaticProps: GetStaticProps = async ({ params }) => {
     const { name: tagName } = await useGetTagInfo.fetchQuery(Number(tagId), queryClient);
 
     // NOTE: tag name 이 api request 값이기 때문에 waterfall 한 fetching 이 필요합니다
-    await prefetchMemesByTag(tagName, queryClient);
+    await useGetMemesByTag.fetchInfiniteQuery(tagName, queryClient);
 
     return {
       props: {
